a truckload of changes all at once, mostly admin perks
admin commands are now permitted based on member permission.

fixed /creepy

updated admin manual

admin perms are checked before running a command

fixed bug with /move and /del error

+++ corePieces/adminCommands.js
@@ -17,16 +17,15 @@ var adminTools = require('./adminTools'),
         channelErr:'Specify the channel of choice within quotation marks (e.g "Rylans png fest")',
         channelNotFound:'*Error: Couldn\'t find any channels!*',
         delMoveBigNum:'Sorry dude, Discord will only let me handle up to 99 messages :/',
-        //@~ - Delete, move; @! - command: del, move
-        delMoveHelp:`Example usage:
-        @~ the last 5 messages:
-        \`/@! 5\`
+        delMoveHelp:(title,command)=>`Example usage:
+        ${title} the last 5 messages:
+        \`${command} 5\`
         
-        @~ messages from @joeSchmoe within a 10 message radius
-        \`/@! 10 @joeSchmoe\`
+        ${title} messages from @joeSchmoe within a 10 message radius
+        \`${command} 10 @joeSchmoe\`
         
-        @~ messages from multiple people within a 50 message radius, as long as they said "pog"
-        \`/@! @joe @caleb "pog" 50\``,
+        ${title} messages from multiple people within a 50 message radius, as long as they said "pog"
+        \`${command} @joe @caleb "pog" 50\``,
         voicePermissions:m=>m.reply("Sorry, In order to do voice channel commands I need the following permissions: Mute Members, Move Members")
     };
 
@@ -59,20 +58,54 @@ function findChannelMacro(m){
     return results;
 }
 
+//All permissions required by lemonbot, when this file executes, there will be a reverse map to speed searches up
+/* After this quick function runs, the object should look something like this:
+{
+    perm1:["MUTE_MEMBERS","MANAGE_MESSAGES"],
+    perm2["MOVE_MEMBERS","SOME_OTHER_RANDOM_PERMISSION"]
+}
+
+It's designed to hold multiple permissions so admin command can be as complex as we want.
+Wouldn't normally build something this way, but it's only used once sooo
+*/
+var permissionsMap = (permsObj=>{
+    var result = {};
+    for(var i in permsObj){
+        for(var j of permsObj[i]){
+            //If the command doesn't exist, create it
+            if(!result[j])
+                result[j] = [];
+
+            //Add permissions to this command if they aren't here yet
+            if(!result[j].length || result[j].indexOf(i) == -1)
+                result[j].push(i);
+        }
+    }
+    return result;
+})({
+    "MUTE_MEMBERS":['mute','umute'],
+    "MOVE_MEMBERS":['raid','voisplit'],
+    "MANAGE_MESSAGES":['del','move']
+});
+
 var commands = {
     'adminhelp':m=>{
-        if(!adminTools.isAdmin(m)) return;
-        
+        //This command is special as it checks for every permission in order to compile a help list relavent to whoever ran it.
         var resultStr = 'Please note that these commands do not have cooldown... **Use responsibly!**\n';
-        for(var i of adminHelpDesc)
-            resultStr+='`\n'+'/'+i[0]+'` - '+i[1];
-        m.channel.send(resultStr);
+        var noPermissions = true;
+        for(var i of adminHelpDesc){
+            if(adminTools.checkPerms(m,permissionsMap[i[0]],false,false)[0]){
+                resultStr+='`\n'+'/'+i[0]+'` - '+i[1];
+                if(noPermissions) noPermissions = false;
+            }
+        }
+        if(noPermissions)
+            m.reply('Sorry, it looks like there aren\'t any admin commands that can be run. Contact the local admin to get some perms!');
+        else m.channel.send(resultStr);
     },
     'del':(m,args)=>{
-        if(!adminTools.isAdmin(m)) return;
-
         if(args.length == 1){
-            m.reply(overUsedVars.delMoveHelp.replaceAll('@!','del').replaceAll('@~','Delete'));
+            m.reply(overUsedVars.delMoveHelp('Delete',args[0]));
             return;
         }
         
@@ -96,11 +129,9 @@ var commands = {
         });
     },
     'move':(m,args)=>{
-        if(!adminTools.isAdmin(m)) return;
-
         //Totally original help page and totally not 100% copied from /del
         if(args.length == 1){
-            m.reply(overUsedVars.delMoveHelp.replaceAll('@!','move').replaceAll('@~','Move'));
+            m.reply(overUsedVars.delMoveHelp('Move',args[0]));
             return;
         }
         
@@ -133,13 +164,34 @@ var commands = {
 
         adminTools.queryMessages(m,currNum,phraseList,messages=>{
             //Copy all messages
-            var finalText = messages.map(e=>'<@'+e.author.id+'> '+e.content).reverse();
-            for(var i of finalText) channelObj.send(i);
-            m.channel.bulkDelete(messages);
+            let finalText = messages.map(e=>{
+                var attachments = [];
+                for(var i of e.attachments){
+                    attachments.push(i[1].attachment);
+                }
+                return '<@'+e.author.id+'> '+(attachments.length?attachments.join('\n')+'\n':'')+e.content;
+            }).reverse();
+            for(let i of finalText){
+                var messagePieces = [''];
+                if(i.length > 2000){
+                    for(var j = 0; j < i.length;j++){
+                        messagePieces[messagePieces.length-1] += i[j];
+                        if(messagePieces[messagePieces.length-1].length == 2000)
+                            messagePieces.push('');
+                    }
+                }
+                else messagePieces[0] = i;
+                for(let j of messagePieces)
+                    channelObj.send(j).then(function(){
+                        //Delete everything if this is the very last item to be moved around [edit - holy cow this is jank XP]
+                        //This is because when an attachment only used once, it is deleted if nothing else needs it. In this way it's a swap of hand from the old message to the new.
+                        if(i == finalText[finalText.length-1] && j == messagePieces[messagePieces.length-1])
+                            m.channel.bulkDelete(messages);
+                    });
+            }
         });
     },
     'mute':(m,args)=>{
-        if(!adminTools.isAdmin(m)) return;
         //First find the voice channel in the args, it should be in quotes
         let muteLimit = 5000*60, //5 minutes
             channels = findChannelMacro(m);
@@ -178,7 +230,6 @@ var commands = {
     'umute':m=>{
         /*This command is special, it can either take a message object or an array of channel objects
         It wouldn't make sense to check for admin upon accepting channel objects so this is avoided*/
-        if(m.content && !adminTools.isAdmin(m)) return;
         //m can be two things, a message or a channel depending on how it was invoked.
         let channels;
         if(!m.content) channels = m;
@@ -207,7 +258,6 @@ var commands = {
         }
     },
     'voisplit':m=>{
-        if(m.content && !adminTools.isAdmin(m)) return;
         //Grab all requested channels
         var channels = findChannelMacro(m);
 
@@ -240,11 +290,9 @@ var commands = {
                 members.pop();
             }
         }
-
     },
 
     'raid':(m,args)=>{
-        if(m.content && !adminTools.isAdmin(m)) return;
         //Grab all requested channels
         var channels = findChannelMacro(m);
 
@@ -275,4 +323,7 @@ var commands = {
     }
 }
 
-module.exports = commands;
+module.exports = {
+    commands:commands,
+    permissionsMap:permissionsMap
+};
